pyproject.toml
[tool.poetry]
name = "django-trainingii"
version = "0.1.0"
description = ""
authors = ["Mohamed Said <muhamdsaid253@gmail.com>"]
readme = "README.md"
packages = [{include = "django_trainingii"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
django-imagekit = "^4.1.0"
pillow = "^9.2.0"
djangorestframework = "^3.14.0"
pygments = "^2.13.0"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
